requirements.txt
faiss-cpu
chromadb
sentence-transformers
numpy
flask
flask-cors
requests
beautifulsoup4
